*Les deux appels (GooglePlace et Firestore) fonctionnent *Mapping partiel et récupération partielle des informations

+++ app/src/main/java/com/example/go4lunch/data/pojo/LatLngLiteralResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class LatLngLiteralResponse {
     @JsonProperty("lat")
-    int lat;
+    Double lat;
     @JsonProperty("lng")
-    int lng;
+    Double lng;
 }
